requirements.txt
# Core Dependencies
boto3==1.34.0
botocore==1.34.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI/ML Libraries
anthropic==0.7.8
langchain==0.1.0
langchain-aws==0.1.0

# Financial Data
alpha-vantage==2.3.1
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3

# Document Generation
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.7
fpdf2==2.7.6

# Data Processing
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AWS Services
aws-cdk-lib==2.110.0
constructs==10.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1